+++ frontend/src/ducks/app/index.js
@@ -1,5 +1,5 @@
 import actions from './actions';
-import types from './types';
+import types, { modalTypes } from './types';
 import selectors from './selectors';
 import reducer, { initialState } from './reducer';
 
@@ -7,6 +7,7 @@ export {
     reducer,
     actions,
     types,
+    modalTypes,
     initialState,
     selectors,
 };