+++ frontend/src/ducks/app/selectors.js
@@ -2,6 +2,9 @@ const selectors = {
     selectScreenWidth(state) {
         return state.app.screenWidth;
     },
+    selectModal(state) {
+        return state.app.modal;
+    }
 };
 
 export default selectors;