+++ frontend/src/ducks/app/types.js
@@ -1,5 +1,13 @@
 const types = {
     SCREEN_RESIZE: '@@app/SCREEN_RESIZE',
+    OPEN_MODAL: '@@app/OPEN_MODAL',
+    HIDE_MODAL: '@@app/HIDE_MODAL',
 };
 
+const modalTypes = {
+    FINISH_RACE: 'FINISH_RACE',
+    CONFIRM_LOGOUT: 'CONFIRM_LOGOUT',
+};
+
+export { modalTypes };
 export default types;